Rename identifier module
